+++ src/1_banner_css.js
@@ -94,6 +94,7 @@ banner_css.css = function(options, element) {
 			style += banner_css.nonie;
 		}
 	}
+	style += options.customCSS;
 
 	if (options.iframe) {
 		style += banner_css.inneriframe;