+++ src/3_branch.js
@@ -1123,8 +1123,9 @@ if (CORDOVA_BUILD) { // jshint undef:false
  * By default, the app banner displays inside of an iFrame (isolates the app banner css from your page), at the top of the page, shows a close button to the user, and will never show again once closed by the user. All of this functionality can be customized.
  * The `iframe` property defaults to true, and can be set to false if you wish for the banner HTML to display within your page. This allows you to customize the CSS of the banner, past what the Web SDK allows.
  * The `disableHide` property defaults to false, and when set to true, removes the close button on the banner.
- * The `forgetHide` property defaults to false, and when set to true, will forget if the user has opened the banner previously, and thus will always show the banner to them even if they have closed it in the past.
- * The `position` property defaults to 'top', but can also be set to 'bottom' if you would prefer to show the app banner from the bottom of the screen.
+ * The `forgetHide` property defaults to false, and when set to true, will forget if the user has opened the banner previously, and thus will always show the banner to them even if they have closed it in the past. It can also be set to an integer, in which case, it would forget that the user has previously hid the banner after X days.
+ * The `position` property, defaults to 'top', but can also be set to 'bottom' if you would prefer to show the app banner from the bottom of the screen.
+ * The `customCSS` property allows you to style the banner, even if it is isolated within an iframe. To assist you with device specific styles, the body element of the banner has one of three classes: `branch-banner-android`, `branch-banner-ios`, or `branch-banner-desktop`.
  * The `mobileSticky` property defaults to false, but can be set to true if you want the user to continue to see the app banner as they scroll.
  * The `desktopSticky` property defaults to true, but can be set to false if you want the user to only see the app banner when they are scrolled to the top of the page.
  * ```js
@@ -1137,10 +1138,11 @@ if (CORDOVA_BUILD) { // jshint undef:false
  *          // Display preferences
  *          iframe: false,
  *          disableHide: true,
- *          forgetHide: true,
+ *          forgetHide: true, // Can also be set to an integer. For example: 10, would forget that the user previously hid the banner after 10 days
  *          position: 'bottom',
  *          mobileSticky: true,
- *          desktopSticky: true
+ *          desktopSticky: true,
+ *          customCSS: '.title { color: #F00; }'
  *     },
  *     {... link data ...}
  * );
@@ -1208,20 +1210,21 @@ if (WEB_BUILD) { // jshint undef:false
  *     icon: 'http://icons.iconarchive.com/icons/wineass/ios7-redesign/512/Appstore-icon.png',
  *     title: 'Branch Demo App',
  *     description: 'The Branch demo app!',
- *     openAppButtonText: 'Open',         // Text to show on button if the user has the app installed
- *     downloadAppButtonText: 'Download', // Text to show on button if the user does not have the app installed
- *     sendLinkText: 'Send Link',         // Text to show on desktop button to allow users to text themselves the app
- *     phonePreviewText: '+44 9999-9999', // The default phone placeholder is a US format number, localize the placeholder number with a custom placeholder with this option
- *     showiOS: true,                     // Should the banner be shown on iOS devices?
- *     showAndroid: true,                 // Should the banner be shown on Android devices?
- *     showDesktop: true,                 // Should the banner be shown on desktop devices?
- *     iframe: true,                      // Show banner in an iframe, recomended to isolate Branch banner CSS
- *     disableHide: false,                // Should the user have the ability to hide the banner? (show's X on left side)
- *     forgetHide: false,                 // Should we remember or forget whether the user hid the banner?
- *     position: 'top',                   // Sets the position of the banner, options are: 'top' or 'bottom', and the default is 'top'
- *     mobileSticky: false,               // Determines whether the mobile banner will be set `position: fixed;` (sticky) or `position: absolute;`, defaults to false *this property only applies when the banner position is 'top'
- *     desktopSticky: true,               // Determines whether the desktop banner will be set `position: fixed;` (sticky) or `position: absolute;`, defaults to true *this property only applies when the banner position is 'top'
- *     make_new_link: false               // Should the banner create a new link, even if a link already exists?
+ *     openAppButtonText: 'Open',              // Text to show on button if the user has the app installed
+ *     downloadAppButtonText: 'Download',      // Text to show on button if the user does not have the app installed
+ *     sendLinkText: 'Send Link',              // Text to show on desktop button to allow users to text themselves the app
+ *     phonePreviewText: '+44 9999-9999',      // The default phone placeholder is a US format number, localize the placeholder number with a custom placeholder with this option
+ *     showiOS: true,                          // Should the banner be shown on iOS devices?
+ *     showAndroid: true,                      // Should the banner be shown on Android devices?
+ *     showDesktop: true,                      // Should the banner be shown on desktop devices?
+ *     iframe: true,                           // Show banner in an iframe, recomended to isolate Branch banner CSS
+ *     disableHide: false,                     // Should the user have the ability to hide the banner? (show's X on left side)
+ *     forgetHide: false,                      // Should we show the banner after the user closes it? Can be set to true, or an integer to show again after X days
+ *     position: 'top',                        // Sets the position of the banner, options are: 'top' or 'bottom', and the default is 'top'
+ *     mobileSticky: false,                    // Determines whether the mobile banner will be set `position: fixed;` (sticky) or `position: absolute;`, defaults to false *this property only applies when the banner position is 'top'
+ *     desktopSticky: true,                    // Determines whether the desktop banner will be set `position: fixed;` (sticky) or `position: absolute;`, defaults to true *this property only applies when the banner position is 'top'
+ *     customCSS: '.title { color: #F00; }',   // Add your own custom styles to the banner that load last, and are gauranteed to take precedence, even if you leave the banner in an iframe
+ *     make_new_link: false                    // Should the banner create a new link, even if a link already exists?
  * }, {
  *     tags: ['tag1', 'tag2'],
  *     feature: 'dashboard',
@@ -1257,6 +1260,7 @@ if (WEB_BUILD) { // jshint undef:false
  *
  */
 	Branch.prototype['banner'] = wrap(callback_params.NO_CALLBACK, function(done, options, data) {
+		if (typeof options['forgetHide'] == 'undefined' && typeof options['forgetHide'] != 'undefined') { options['forgetHide'] = options['forgetHide']; }
 		var bannerOptions = {
 			icon: options['icon'] || '',
 			title: options['title'] || '',
@@ -1270,11 +1274,12 @@ if (WEB_BUILD) { // jshint undef:false
 			showAndroid: typeof options['showAndroid'] == 'undefined' ? true : options['showAndroid'],
 			showDesktop: typeof options['showDesktop'] == 'undefined' ? true : options['showDesktop'],
 			disableHide: !!options['disableHide'],
-			forgetHide: !!options['forgetHide'],
+			forgetHide: typeof options['forgetHide'] == 'number' ? options['forgetHide'] : !!options['forgetHide'],
 			position: options['position'] || 'top',
-			make_new_link: !!options['make_new_link'],
+			customCSS: options['customCSS'] || '',
 			mobileSticky: typeof options['mobileSticky'] == 'undefined' ? false : options['mobileSticky'],
-			desktopSticky: typeof options['desktopSticky'] == 'undefined' ? true : options['desktopSticky']
+			desktopSticky: typeof options['desktopSticky'] == 'undefined' ? true : options['desktopSticky'],
+			make_new_link: !!options['make_new_link']
 		};
 		if (typeof options['showMobile'] != 'undefined') {
 			bannerOptions.showiOS = bannerOptions.showAndroid = options['showMobile'];